--- src/main/java/com/moyo/backend/follow/application/FollowCommandRepository.java
@@ -1,4 +1,0 @@
-package com.moyo.backend.follow.application;
-
-public interface FollowCommandRepository {
-}
