--- src/main/java/com/moyo/backend/follow/application/FollowQueryRepository.java
@@ -1,6 +1,0 @@
-package com.moyo.backend.follow.application;
-
-public interface FollowQueryRepository {
-
-
-}
