+++ src/main/java/com/moyo/backend/follow/application/FollowRepository.java
@@ -4,7 +4,11 @@ import com.moyo.backend.follow.dto.GithubUserResponse;
 
 import java.util.List;
 
-public interface GithubFollowQueryClient {
+public interface FollowRepository {
+
+    int follow(String username, String accessToken);
+
+    int unfollow(String username, String accessToken);
 
     List<GithubUserResponse> getFollowingList(String accessToken);
