--- src/main/java/com/moyo/backend/follow/application/GithubFollowCommandClient.java
@@ -1,8 +1,0 @@
-package com.moyo.backend.follow.application;
-
-public interface GithubFollowCommandClient {
-
-    int follow(String username, String accessToken);
-
-    int unfollow(String username, String accessToken);
-}
