+++ src/main/java/com/moyo/backend/follow/infrastructure/httpClient/FollowHttpClient.java
@@ -1,6 +1,6 @@
-package com.moyo.backend.follow.infrastructure.github;
+package com.moyo.backend.follow.infrastructure.httpClient;
 
-import com.moyo.backend.follow.application.GithubFollowQueryClient;
+import com.moyo.backend.follow.application.FollowRepository;
 import com.moyo.backend.follow.dto.GithubUserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.ParameterizedTypeReference;
@@ -10,11 +10,44 @@ import org.springframework.web.client.RestClient;
 
 import java.util.List;
 
+
 @Component
 @RequiredArgsConstructor
-public class GithubFollowQueryRestApiClient implements GithubFollowQueryClient {
+public class FollowHttpClient implements FollowRepository {
 
     private final RestClient restClient;
+
+    @Override
+    public int follow(String username, String accessToken){
+        return restClient.put()
+                .uri("https://api.github.com/user/following/"+username)
+                .headers(
+                        header ->{
+                            header.setBearerAuth(accessToken);
+                            header.set(HttpHeaders.ACCEPT, "application/vnd.github+json");
+                            header.set("X-GitHub-Api-Version", "2022-11-28");
+                        }
+                )
+                .retrieve()
+                .toBodilessEntity()
+                .getStatusCode().value();
+    }
+
+    @Override
+    public int unfollow(String username, String accessToken){
+        return restClient.delete()
+                .uri("https://api.github.com/user/following/"+username)
+                .headers(
+                        header ->{
+                            header.setBearerAuth(accessToken);
+                            header.set(HttpHeaders.ACCEPT, "application/vnd.github+json");
+                            header.set("X-GitHub-Api-Version", "2022-11-28");
+                        }
+                )
+                .retrieve()
+                .toBodilessEntity()
+                .getStatusCode().value();
+    }
 
     @Override
     public List<GithubUserResponse> getFollowingList(String accessToken){
@@ -47,4 +80,5 @@ public class GithubFollowQueryRestApiClient implements GithubFollowQueryClient {
                 .retrieve()
                 .body(new ParameterizedTypeReference<>() {});
     }
+
 }
