+++ src/main/java/com/moyo/backend/follow/ui/FollowController.java
@@ -1,8 +1,7 @@
 package com.moyo.backend.follow.ui;
 
 import com.moyo.backend.common.model.ApiResponse;
-import com.moyo.backend.follow.application.FollowCommandService;
-import com.moyo.backend.follow.application.FollowQueryService;
+import com.moyo.backend.follow.application.FollowService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -15,15 +14,14 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class FollowController {
 
-    private final FollowQueryService followQueryService;
-    private final FollowCommandService followCommandService;
+    private final FollowService followService;
 
     @PostMapping("/follow/{username}")
     public ResponseEntity<ApiResponse<Void>> follow(@PathVariable("username") String username, @AuthenticationPrincipal GithubOAuth2User userPrincipal){
 
         log.info("[팔로우 요청] | 요청자 ProviderId: {} -> 대상: {}", userPrincipal.getName(), username);
 
-        followCommandService.follow(username, userPrincipal);
+        followService.follow(username, userPrincipal);
 
         return ResponseEntity.ok(ApiResponse.noContent());
     }
@@ -33,7 +31,7 @@ public class FollowController {
 
         log.info("[언 팔로우 요청] | 요청자 ProviderId: {} -> 대상: {}", userPrincipal.getName(), username);
 
-        followCommandService.unfollow(username, userPrincipal);
+        followService.unfollow(username, userPrincipal);
 
         return ResponseEntity.ok(ApiResponse.noContent());
     }
@@ -42,21 +40,21 @@ public class FollowController {
     @GetMapping("/users/me/followings/only")
     public ResponseEntity<?> getFollowingOnlyList(@AuthenticationPrincipal GithubOAuth2User userPrincipal, Pageable pageable){
 
-        return ResponseEntity.ok(ApiResponse.success(followQueryService.getFollowingOnlyList(userPrincipal)));
+        return ResponseEntity.ok(ApiResponse.success(followService.getFollowingOnlyList(userPrincipal)));
     }
 
     // 상대만 나를 팔로잉
     @GetMapping("/users/me/followers/only")
     public ResponseEntity<?> getFollowerOnlyList(@AuthenticationPrincipal GithubOAuth2User userPrincipal, Pageable pageable){
 
-        return ResponseEntity.ok(ApiResponse.success(followQueryService.getFollowerOnlyList(userPrincipal)));
+        return ResponseEntity.ok(ApiResponse.success(followService.getFollowerOnlyList(userPrincipal)));
     }
 
     // 맞 팔로우
     @GetMapping("/users/me/followings/mutual")
     public ResponseEntity<?> getMutualFollowingList(@AuthenticationPrincipal GithubOAuth2User userPrincipal, Pageable pageable){
 
-        return ResponseEntity.ok(ApiResponse.success(followQueryService.getMutualFollowList(userPrincipal)));
+        return ResponseEntity.ok(ApiResponse.success(followService.getMutualFollowList(userPrincipal)));
     }
 
 }
